Moved /submitcards route in expressServer.js to its own router file
