+++ expressServer.js
@@ -7,6 +7,7 @@ const compression = require('compression')
 const userRouter = require('./routers/users')
 const { router: authRouter, isAuth, hasSecurityQuestion } = require('./routers/auth')
 const settingRouter = require('./routers/settings')
+const submitcardsRouter = require('./routers/submitcards')
 const collectionsModel = require('./models/collections')
 const OpenAI = require('openai')
 const openai = new OpenAI({
@@ -66,11 +67,10 @@ app.use('/securityQuestions', securityQuestionsRouter)
 app.use('/collection', isAuth, hasSecurityQuestion, collectionRouter)
 app.use('/check', isAuth, hasSecurityQuestion)
 app.use('/review', isAuth, hasSecurityQuestion)
-app.use('/submitcards', isAuth, hasSecurityQuestion)
+app.use('/submitcards', isAuth, hasSecurityQuestion, submitcardsRouter)
 app.use('/generate', isAuth, hasSecurityQuestion)
 app.use('/api/generate', isAuth, hasSecurityQuestion)
 app.use('/home', isAuth, hasSecurityQuestion, homeRouter)
-
 app.get('/health', (_, res) => {
     return res.status(200).send('ok')
 })
@@ -204,51 +204,6 @@ app.get('/check', (req, res) => {
     const carouselData = { bg: '/images/plain-FFFFFF.svg', cards: data, queryType: 'finalize', pictureID: req.session.picture }
 
     return res.render('review', carouselData)
-})
-
-app.post('/submitcards', async (req, res) => {
-    let lastShareCode
-    let shareId
-
-    // get the latest sharecode from collections
-    try {
-        const result = await collectionsModel.findOne().sort({ shareId: -1 }).select('shareId').exec()
-        lastShareCode = result ? result.shareId : null
-    } catch (err) {
-        console.log('Failed to fetch latestShareCode')
-    }
-
-    if (lastShareCode === null) {
-        shareId = 0
-    } else {
-        shareId = lastShareCode + 1
-    }
-
-    const inputData = JSON.parse(req.body.cards).map((card) => {
-        return {
-            shareId: `${shareId}`,
-            ...card,
-        }
-    })
-
-    const transactionSession = await mongoose.startSession()
-    transactionSession.startTransaction()
-    try {
-        await flashcardsModel.insertMany(inputData, { session: transactionSession })
-        console.log('flashcards insert ok')
-        await collectionsModel.create([{ setName: `${req.body.name}`, userId: req.session.userId, shareId: shareId }], { session: transactionSession })
-        console.log('set insert ok')
-        await transactionSession.commitTransaction()
-        transactionSession.endSession()
-        console.log(`Successfully wrote ${req.body.name} to db`)
-    } catch (err) {
-        await transactionSession.abortTransaction()
-        transactionSession.endSession()
-        console.log('Error inserting db')
-    }
-
-    res.status(200)
-    res.json(JSON.stringify({ shareId: shareId }))
 })
 
 app.get('/egg', (req, res) => {
